+++ src/jormungand/api/__init__.py
@@ -1,7 +1,7 @@
 from datamodel import DataModelPluginInterface
 from extraction import ExtractionPluginInterface
 from postprocessing import PostProcessingPluginInterface
-from validation import ValidationPluginInterface
 from storage import StoragePluginInterface
+from validation import ValidationPluginInterface
 
 __author__ = 'adam.jorgensen.za@gmail.com'
