+++ src/jormungand/plugins/storage/StatisticsStorage.py
@@ -1,5 +1,6 @@
 import os
-from jormungand.api import StoragePluginInterface
+
+from src.jormungand.api import StoragePluginInterface
 
 __author__ = 'adam.jorgensen.za@gmail.com'
